// src/shared/enums/user-activity.enum.ts
export enum UserActionEnum {
  REGISTER = "REGISTER",
  LOGIN = "LOGIN",
  LOGOUT = "LOGOUT",
  FORGOT_PASSWORD = "FORGOT_PASSWORD",
  RESET_PASSWORD = "RESET_PASSWORD",
  EMAIL_VERIFIED = "EMAIL_VERIFIED",

  PROFILE_UPDATED = "PROFILE_UPDATED",

  APPOINTMENT_CREATED = "APPOINTMENT_CREATED",
  APPOINTMENT_CANCELLED = "APPOINTMENT_CANCELLED",
}